==> 01-Avoiding-Data-Leakage.ipynb <==
# Data Leakage and its Avoidance

If our workflow is:
1. Prepare the data set (eliminate redundancy, impute, scale, ...)
2. Split the data set
3. Evaluate the model
then information from the test set contaminates the preparation and fitting of the training data

To avoid this the workflow should be:
1. Split the data set
2. **Prepare using the training data set**
3. **Apply the same preparation to the Training and Test data sets**
4. Evaluate the model


from sklearn.datasets import make_classification
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler

from sklearn.metrics import mean_squared_error
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import Ridge, RidgeCV, Lasso, LassoCV
from sklearn.feature_selection import SequentialFeatureSelector

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.metrics import accuracy_score
import pandas as pd

df = pd.read_csv('wildfires.csv')
df.head(2)

#drop first two index columns
df = df.iloc[:, 2:]

#check for nulls
df.isna().sum()

#just rearranging columns so fire-size is first (y)
cols = list(df.columns)
cols = [cols[1]] + cols[0:1] + cols[2:]
df = df[cols]

df.columns

df.shape

#encode stat_cause_descr
df["cause_code"] = df["stat_cause_descr"].astype('category').cat.codes

c = df["stat_cause_descr"].astype('category')
cause_dict = dict(enumerate(c.cat.categories))
print(cause_dict)

#encode state
df["state_code"] = df["state"].astype('category').cat.codes

s = df["state"].astype('category')
state_dict = dict(enumerate(s.cat.categories))
print(state_dict)


#encode fire size class
df["fire_size_class_code"] = df["fire_size_class"].astype('category').cat.codes

s = df["fire_size_class"].astype('category')
fire_size_class_dict = dict(enumerate(s.cat.categories))

print(fire_size_class_dict)

#encode discovery month
month_dict = {'Jan':1, 'Feb':2, 'Mar':3, 'Apr':4, 'May':5, 'Jun':6, 'Jul':7, 'Aug':8, 'Sep':9, 'Oct':10, 'Nov':11, 'Dec':12}

df.discovery_month = df.discovery_month.map(month_dict)

df.head(2)

#export clean df
df.to_csv('wildfires_clean.csv', index=False)

pd.set_option('display.max_columns', df.shape[1])
print(df.info())

Index(['fire_size', 'fire_name', 'fire_size_class', 'stat_cause_descr',
       'latitude', 'longitude', 'state', 'disc_clean_date', 'cont_clean_date',
       'discovery_month', 'disc_date_final', 'cont_date_final', 'putout_time',
       'disc_date_pre', 'disc_pre_year', 'disc_pre_month', 'wstation_usaf',
       'dstation_m', 'wstation_wban', 'wstation_byear', 'wstation_eyear',
       'Vegetation', 'fire_mag', 'weather_file', 'Temp_pre_30', 'Temp_pre_15',
       'Temp_pre_7', 'Temp_cont', 'Wind_pre_30', 'Wind_pre_15', 'Wind_pre_7',
       'Wind_cont', 'Hum_pre_30', 'Hum_pre_15', 'Hum_pre_7', 'Hum_cont',
       'Prec_pre_30', 'Prec_pre_15', 'Prec_pre_7', 'Prec_cont', 'remoteness'],
      dtype='object')

X = df[['cause_code', 'state_code', 'fire_size_class_code', 'discovery_month','Vegetation',
       'fire_mag', 'Temp_pre_30', 'Temp_pre_15', 'Temp_pre_7',
       'Temp_cont', 'Wind_pre_30', 'Wind_pre_15', 'Wind_pre_7', 'Wind_cont',
       'Hum_pre_30', 'Hum_pre_15', 'Hum_pre_7', 'Hum_cont', 'Prec_pre_30',
       'Prec_pre_15', 'Prec_pre_7', 'Prec_cont', 'remoteness']]

y = df[['fire_size']]


print(X.shape)
print(y.shape)

### Correct workflow avoiding data leakage

X.info()

y.info()

# split first
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=1)

scaler = MinMaxScaler()
scaler.fit(X_train)
X_train = scaler.transform(X_train)

# scale the test dataset
X_test = scaler.transform(X_test)
model1 = LinearRegression()

model1.fit(X_train, y_train)

yhat = model1.predict(X_test)

print("Train R-squared: ", model1.score(X_train,y_train))
print('Train MSE: %.2f'% mean_squared_error(y_train, model1.predict(X_train)))
print("Testing R-squared: ", model1.score(X_test,y_test))
print('Test MSE: %.2f'% mean_squared_error(y_test, model1.predict(X_test)))

### Use a Pipeline to avoid data leakage during k-fold CV

from sklearn.pipeline import Pipeline

X, y = make_classification(n_samples=55367, n_features=21, 
                           n_informative=15, n_redundant=5, 
                           random_state=7)

pipe = Pipeline(steps = 
                [('scaler', MinMaxScaler()),
                 ('model', LinearRegression())])

cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=1)

scores = cross_val_score(pipe, X, y, scoring='neg_mean_squared_error', cv=cv, n_jobs=-1)

print('Accuracy: %.3f (%.3f)' % (-scores.mean(), scores.std()))
