==> 02-Sparse-Column-Identification.ipynb <==
# Identifying "sparse columns" --- columns without predictive information

We use the following data set
https://www.kaggle.com/datasets/ashrafkhan94/oil-spill

import pandas as pd
df = pd.read_csv('wildfires_clean.csv')

df.head(2)

df.columns

df.shape

### Count unique values in each column

df.nunique(axis=0)#.head(10)

df.columns[df.nunique() == 1]

## Calculate the percentage of unique values in each column

df.nunique()/len(df)*100

## Columns with less than 1% unique values

list(df.columns[df.nunique()/len(df) < 0.01])

## Using Variance Threshold

$s^2 = \frac{\sum_1^N(x_i - \mu)}{N-1}$

df.head()

df2 = df[['cause_code', 'state_code', 'fire_size_class_code', 'Vegetation',
       'fire_mag', 'Temp_pre_30', 'Temp_pre_15', 'Temp_pre_7',
       'Temp_cont', 'Wind_pre_30', 'Wind_pre_15', 'Wind_pre_7', 'Wind_cont',
       'Hum_pre_30', 'Hum_pre_15', 'Hum_pre_7', 'Hum_cont', 'Prec_pre_30',
       'Prec_pre_15', 'Prec_pre_7', 'Prec_cont', 'remoteness']]

# # y = df.iloc[:,1: ]
# y = df.iloc[:, 1:]
# y

from sklearn.feature_selection import VarianceThreshold

data = df2.values

X = data[:, 1:] #everything except the first column
y = data[:, 0] #first column

print(X.shape, y.shape)

transform = VarianceThreshold()

X_sel = transform.fit_transform(X)
print(X_sel.shape)

df2.columns[:-1][~transform.get_support()]

transform = VarianceThreshold(threshold=0.05)
transform.fit(X)

[i for i, b in enumerate(transform.get_support()) if not b]

df.columns[:-1][~transform.get_support()]

from numpy import arange
import altair as alt

# df = pd.read_csv('oil-spill.csv', header=None)

# data = df.values
# X = data[:, :-1]
# y = data[:, -1]

# print(X.shape, y.shape)

thresholds = arange(0.0, 0.55, 0.05)

results = []
for t in thresholds:
    
    transform = VarianceThreshold(threshold=t)
    
    X_sel = transform.fit_transform(X)
    rows, cols = X_sel.shape
    n_features = cols
    print('Threshold=%.2f, Features=%d' % (t, n_features))
    
    results.append(n_features)
    
d2 = pd.DataFrame({'threshold': thresholds, 'n_features': results})
alt.Chart(d2).mark_line().encode(
    x='threshold',
    y='n_features')
