==> 05-Handling-Missing-Values.ipynb <==
# Handling Missing Values
https://www.kaggle.com/datasets/uciml/pima-indians-diabetes-database

import pandas as pd

df = pd.read_csv('wildfires_clean.csv', header=0)
# summarize the dataset
df.head(10)

df.info()

features = df.columns[1:]
features

num_missing = (df[features] == 0).sum()
# report the results
print(num_missing)

import numpy as np

df[features] = df[features].replace(0, np.nan)
# count the number of nan values in each column
df.isnull().sum()

df.head(20)

df.shape


df1=df

nan_percentages = (df1.isna().sum(axis=1) / len(df1.columns)) * 100
nan_percentages

#df.dropna(inplace=True)
#(3507, 43)


threshold = 50
df1.drop(nan_percentages[nan_percentages > threshold].index, inplace=True)
df1.shape

df.head(20)
