==> 07-Imputation-and-evaluation.ipynb <==
# Imputing missing values
https://www.kaggle.com/datasets/uciml/horse-colic

https://archive.ics.uci.edu/ml/datasets/Horse+Colic

The features from the dataset that we are planning not to use contain the missing values and the features that we will use does not contain missing values. Hence, we do not have the need to use imputation right now.

import pandas as pd
import numpy as np

df = pd.read_csv('horse-colic.csv', header=None,)
df.head().T

When the data is read missing values of `?` can be replaced with `NaN`

df = pd.read_csv('horse-colic.csv', header=None, na_values='?')
df.head()

df.isna().sum()

Determining the percentage of missing values

pd.concat([df.isna().sum(), df.isna().sum()/len(df)*100], axis=1).head(10)

target = 23
features = list(df.columns)
features.remove(target)

X = df[features]
y = df[target]

from sklearn.impute import SimpleImputer
imputer = SimpleImputer(strategy='mean')
imputer.fit(X)

Ximputed = imputer.transform(X)

sum(np.isnan(Ximputed).flatten())

pd.DataFrame(Ximputed, index=X.index, columns=X.columns)#.isna().sum()

# What is wrong with the above?

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.pipeline import Pipeline

model = RandomForestClassifier()
imputer = SimpleImputer(strategy='mean')

pipeline = Pipeline(steps=
                    [('i', imputer), 
                     ('m', model)])

# define model evaluation
cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=1)

# evaluate model
scores = cross_val_score(pipeline, X, y, scoring='accuracy', cv=cv, n_jobs=-1)

print('Mean Accuracy: %.3f (%.3f)' % (scores.mean(), scores.std()))

results = []
strategies = ['mean', 'median', 'most_frequent'] #, 'constant']
for s in strategies:

    pipeline = Pipeline(steps=
                        [('i', SimpleImputer(strategy=s)), 
                         ('m', RandomForestClassifier())])
    
    cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=1)
    scores = cross_val_score(pipeline, X, y, scoring='accuracy', cv=cv, n_jobs=-1)

    results.append(scores)
    print('%20s %.3f (%.3f)' % (s, scores.mean(), scores.std()))

## Using a Knn imputer

from sklearn.impute import KNNImputer

results = []
nbrhood_size = [1,3,5,7,9,15,18,21]
for k in nbrhood_size:
    # create the modeling pipeline
    pipeline = Pipeline(steps=
                        [('i', KNNImputer(n_neighbors=k)), 
                         ('m', RandomForestClassifier())])
    # evaluate the model
    cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=1)
    scores = cross_val_score(pipeline, X, y, scoring='accuracy', cv=cv, n_jobs=-1)
    # store results
    results.append(scores)
    print('k=%3d %.3f (%.3f)' % (k, scores.mean(), scores.std()))


## MICE imputation

from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer

model = RandomForestClassifier()
imputer = IterativeImputer()
pipeline = Pipeline(steps=
                    [('i', imputer), 
                     ('m', model)])

cv = RepeatedStratifiedKFold(n_splits=10, n_repeats=3, random_state=1)

scores = cross_val_score(pipeline, X, y, scoring='accuracy', cv=cv, n_jobs=-1)
print('Mean Accuracy: %.3f (%.3f)' % (scores.mean(), scores.std()))
